requirements.txt
amqp==5.0.6
appnope==0.1.2
APScheduler==3.6.3
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.8
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-editorjs-fields==0.2.5
django-timezone-field==4.2.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.3.0
idna==3.3
ipython==7.29.0
jedi==0.18.0
kombu==5.1.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.21
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
sentry-sdk
